add group users service and fix message service
