+++ lib/services/message/messages.service.dart
@@ -11,6 +11,24 @@ class MessagesService {
         'Authorization': token,
       },
     );
+    print(response.body);
+
+    if (response.statusCode == 200) {
+      var res = json.decode(response.body);
+
+      return Messages.fromJson(res);
+    } else {
+      throw Exception('Failed to load messages');
+    }
+  }
+
+  Future<Messages> getLastMessageTwoUsers(String token, String userId) async {
+    final response = await http.get(
+      Uri.parse('${dotenv.env['API_URL']}/users/$userId/message'),
+      headers: <String, String>{
+        'Authorization': token,
+      },
+    );
 
     if (response.statusCode == 200) {
       var res = json.decode(response.body);
